Turned on coffeebeans gem and moved coffeescript to production #5
The coffeebeans gem to allow coffeescript to be run in the html.erb
views using the helper tags coffee_script_tag and coffee_script.

Also moved coffeescript and sass to global include so that the app
will work when deployed on heroku. This is needed for the time being
to get the alpha online.