Fixed issue with creating a booking  #12
Fixed an issue where you could not create a booking due to the num
spots available check. Now we check to ensure that the num_spots
param is not nil and that the people list are not nil before trying
to calculate. When a booking is new these will be nil and we just set
the number available to 0.