Updated CSS for the site #15
Updated CSS to change the header and make it fixed at the top of the
page. It will stay there even when scrolling.

Also updated the h1 elements at the top of each page to add padding
and center the headers.

Also updated the faq css to better spacing and no bullets. FAQs are
also now centered on the page.